fix: refactoring modal component
