+++ components/DeleteComponent/index.tsx
@@ -1,18 +1,16 @@
 import TrashButton from '../../public/trash.svg';
 import ColorfulGarbage from '../../public/colorfulGarbage.svg';
 import { Button } from '../Button';
-import { ModalTrigger } from '../ModalTrigger';
-import { ModalContent } from '../ModalContent';
+import { Modal } from '../Modal';
 
 export default function DeleteComponent () {
     return (
-      <ModalTrigger button={<TrashButton />}>
-        <ModalContent
+        <Modal
+          trigger={<TrashButton />}
           icon={<ColorfulGarbage />}
           buttonConfirm={<Button variant='filled'>Com certeza</Button>}
           buttonCancel={<Button variant='ghost'>Não</Button>}
           description='Quer excluir suas aulas?'
         />
-      </ModalTrigger>
     );
 }
